As per VMMaker.oscog-eem.975xi, fFix 64-bit issues in BitBlt Drop MIDI and RE plugins.
git-svn-id: http://squeakvm.org/svn/squeak/trunk@3171 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: 8cabf8b1ae1038517ed6c53ae56150547de23ee2

+++ platforms/Cross/plugins/AsynchFilePlugin/AsynchFilePlugin.h
@@ -10,9 +10,9 @@ typedef struct {
 } AsyncFile;
 
 int asyncFileClose(AsyncFile *f);
-int asyncFileOpen(AsyncFile *f, int fileNamePtr, int fileNameSize, int writeFlag, int semaIndex);
+int asyncFileOpen(AsyncFile *f, long fileNamePtr, int fileNameSize, int writeFlag, int semaIndex);
 int asyncFileRecordSize();
-int asyncFileReadResult(AsyncFile *f, int bufferPtr, int bufferSize);
+int asyncFileReadResult(AsyncFile *f, long bufferPtr, int bufferSize);
 int asyncFileReadStart(AsyncFile *f, int fPosition, int count);
 int asyncFileWriteResult(AsyncFile *f);
-int asyncFileWriteStart(AsyncFile *f, int fPosition, int bufferPtr, int bufferSize);
+int asyncFileWriteStart(AsyncFile *f, int fPosition, long bufferPtr, int bufferSize);
